重构第一版

+++ src/main/java/com/yaxin/userCenter/service/impl/user/DeleteUserService.java
@@ -1,5 +1,6 @@
 package com.yaxin.userCenter.service.impl.user;
 
+import com.yaxin.userCenter.common.Encode;
 import com.yaxin.userCenter.common.Validate;
 import com.yaxin.userCenter.model.BeanFactory;
 import com.yaxin.userCenter.model.item.User;
@@ -7,8 +8,6 @@ import com.yaxin.userCenter.service.BaseService;
 import net.sf.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.Map;
 
 /**
  * 删除用户业务类
@@ -20,19 +19,23 @@ public class DeleteUserService extends BaseService {
     BeanFactory beanFactory;
 
     @Override
-    protected void validate(Map<String, Object> map) throws Exception {
+    protected void before(JSONObject param) throws Exception {
         //本例只做了基本的判断，实际业务中，可能需要更多的业务判断，
-        if(!Validate.existByJson(map, "id")) {
-            throw new Exception("传入的参数有误.请检查传入的参数. " + map);
-        }
+        Validate.existByJson(param, "token","loginname");
+        //解码
+        param.put("loginname", Encode.base64Decode(String.valueOf(param.get("loginname"))));
+        //token校验
     }
 
     @Override
-    protected Result handle(String req) throws Exception {
-        Map<String, Object> map = (Map<String, Object>) JSONObject.toBean(JSONObject.fromObject(req), Map.class);
+    protected Result handle(JSONObject param) throws Exception {
+        if(!"1".equals(param.get("token")))
+            return new Result().failed("未授权用户。通关秘籍 试试token = 1");
         //真正的删除业务会加入很多业务方面的判断，这里只实现了最简单的功能
         User user=beanFactory.getUserItem();
-        user.setId(Integer.valueOf(String.valueOf(map.get("id"))));
+        user.setLoginname(String.valueOf(param.get("loginname")));
+        if(user.findByName() == null)
+            return new Result().failed("没有找到用户.");
         user.delete();
         return new Result().succeeded("OK");
     }
